requirements.txt
apns2==0.3.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.1
daphne==1.3.0
defusedxml==0.5.0
Django==1.11.6
django-braces==1.11.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-nested-admin==3.0.21
django-oauth-toolkit==1.0.0
django-push-notifications==1.5.0
django-redis-cache==1.7.1
django-rest-swagger==2.1.2
django-stdimage==3.0.0
djangorestframework==3.7.1
environ==1.0
google-auth==1.2.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.9.0
msgpack-python==0.4.8
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
pkg-resources==0.0.0
priority==1.3.0
progressbar2==3.34.3
psycopg2==2.7.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.5.3
pymongo==2.8
pyOpenSSL==17.3.0
python-monkey-business==1.0.0
python-utils==2.2.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
service-identity==17.0.0
simplejson==3.12.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
tornado==4.5.2
Twisted==17.9.0
txaio==2.8.2
uritemplate==3.0.0
urllib3==1.22
zope.interface==4.4.3
